+++ compositor_common/src/util/degree.rs
@@ -1,0 +1,4 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone, Default, PartialEq, Eq)]
+pub struct Degree(pub i32);
